change the test for blanks in string
